// hotel-reservation-site-angular/src/app/new-hotel-interface.ts
export interface NewHotelInterface {
    hotel_id: number;
    hotel_name: string;
    hotel_city: string;
    hotel_location: string;
    hotel_stars: string;
    hotel_distance: string;
    hotel_phone: string;
    hotel_price_min: string;
    hotel_price_max: string;
    hotel_avg_rate: string;
    hotel_availability: string;
}